+++ composeApp/src/commonMain/kotlin/echo/music/enhanced/di/ViewModelModule.kt
@@ -109,6 +109,7 @@ val viewModelModule =
                 get(),
                 get(),
                 get(),
+                get(),
             )
         }
         viewModel {